requirements.txt
alembic
annotated-doc
annotated-types
anyio
asyncpg
bcrypt
black
certifi
click
colorama
dnspython
email-validator
fastapi
greenlet
h11
httpcore
httptools
httpx
idna
iniconfig
itsdangerous
Mako
MarkupSafe
mypy
mypy_extensions
packaging
passlib
pathspec
platformdirs
pluggy
pydantic
pydantic_core
Pygments
PyJWT
pytest
pytest-asyncio
python-dotenv
python-telegram-bot
pytokens
pytz
PyYAML
ruff
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets